wip: - Moved tests from middleware over to controller
